Move create link header logic to utility
