+++ src/cashier_frontend_new/src/modules/creationLink/components/createLinkHeader.svelte
@@ -5,7 +5,7 @@
     CREATE_LINK_PROGRESS_SEGMENTS,
     getCreateLinkCardHeaderDisplayName,
     getCreateLinkProgress,
-  } from "$modules/creationLink/services/createLinkHeader";
+  } from "$modules/creationLink/utils/createLinkHeader";
   import { LinkStep } from "$modules/links/types/linkStep";
   import { ChevronLeft } from "lucide-svelte";
