+++ src/cashier_frontend_new/src/modules/creationLink/utils/createLinkHeader.ts
